requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
gotrue==2.12.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
postgrest==1.1.1
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.5.3
ruff==0.12.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
websockets==15.0.1
